refactor: split daemon.rs (1317 lines) into per-platform modules (issue #505)
拆分动机（issue #505）：
- 老的 daemon.rs 单文件 1317 行，三平台实现混在一起：
  - macOS launchd (~300 行)
  - Linux systemd + detached redeploy (~770 行)
  - Windows Task Scheduler (~220 行)
  - 共享路径 helper (~30 行)
- 新人 review 时难以建立全局心智模型，merge 冲突频发，
  dead code 容易藏在大文件里

拆分方案：
- daemon/mod.rs: DaemonAction enum + handle_daemon_command dispatcher
- daemon/common.rs: 跨平台路径定位 helper (ntd_binary_path / ntd_dir / ntd_bin_dir)
- daemon/macos.rs: launchd 实现，仅 #[cfg(target_os = "macos")] 编译
- daemon/linux.rs: systemd 实现，仅 #[cfg(target_os = "linux")] 编译
- daemon/windows.rs: Task Scheduler 实现，仅 #[cfg(target_os = "windows")] 编译
- daemon/redeploy.rs: Linux detached redeploy (DaemonInstallMode /
  RedeployCommandSpec / RedeployError / spawn_detached_redeploy) +
  check_linger helper

Public 表面保持完全兼容：
- ntd::daemon::DaemonAction - clap 子命令枚举
- ntd::daemon::handle_daemon_command(action) - 入口 dispatcher
- ntd::daemon::spawn_detached_redeploy(script) - 在 daemon 模块根 re-export
- ntd::daemon::build_redeploy_spec / DaemonInstallMode / RedeployCommandSpec /
  RedeployError - 全部从 redeploy 模块 re-export，保留旧路径

新单测覆盖：
- daemon::common::tests::test_ntd_binary_path_returns_absolute
- daemon::common::tests::test_ntd_dir_falls_back_to_tmp_when_home_missing
- daemon::common::tests::test_ntd_bin_dir_is_parent_of_binary
- daemon::tests::test_daemon_action_variants_are_exhaustive
  (用 match 全覆盖防止未来重构误删 DaemonAction 变体)

保留所有原注释（per CLAUDE.md 的逐行注释规范），
doc comment / inline comment 都没动，只是把内容搬到对应平台模块下。

Fixes #505
